# entregaveis/openapi.json
{"openapi":"3.1.0","info":{"title":"Market Pulse  Demo API","description":"API de consulta de artigos (demo local).","version":"1.0.0"},"paths":{"/api/articles":{"get":{"summary":"List Articles","operationId":"list_articles_api_articles_get","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":1,"title":"Page"}},{"name":"per_page","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":10,"title":"Per Page"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Page"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/articles/search":{"get":{"summary":"Search Articles","operationId":"search_articles_api_articles_search_get","parameters":[{"name":"q","in":"query","required":true,"schema":{"type":"string","minLength":2,"title":"Q"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","default":1,"title":"Page"}},{"name":"per_page","in":"query","required":false,"schema":{"type":"integer","default":10,"title":"Per Page"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Page"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/articles/{article_id}":{"get":{"summary":"Get Article","operationId":"get_article_api_articles__article_id__get","parameters":[{"name":"article_id","in":"path","required":true,"schema":{"type":"string","title":"Article Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Article"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Article":{"properties":{"id":{"type":"string","title":"Id"},"title":{"type":"string","title":"Title"},"content":{"type":"string","title":"Content"},"published_at":{"type":"string","title":"Published At"},"source":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Source"},"tickers":{"items":{"type":"string"},"type":"array","title":"Tickers","default":[]},"topics":{"items":{"type":"string"},"type":"array","title":"Topics","default":[]},"sentiment":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Sentiment"}},"type":"object","required":["id","title","content","published_at"],"title":"Article"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Page":{"properties":{"items":{"items":{"$ref":"#/components/schemas/Article"},"type":"array","title":"Items"},"page":{"type":"integer","title":"Page"},"per_page":{"type":"integer","title":"Per Page"},"total":{"type":"integer","title":"Total"}},"type":"object","required":["items","page","per_page","total"],"title":"Page"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}